Return refreshed tokens when refreshing
